+++ caim_base/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
-from .models import User
+from .models.awg import User
 
 
 class NewUserForm(UserCreationForm):
