+++ caim_base/views/shortlist.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import BadRequest
 from django.http import JsonResponse
-from ..models import Animal, AnimalShortList
+from ..models.animals import Animal, AnimalShortList
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_http_methods
